% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_import_field_notes.R
\name{import_field_notes}
\alias{import_field_notes}
\alias{import_fieldnotes}
\title{Import field notes from an Excel spreadsheet}
\usage{
import_field_notes(path, ...)

import_fieldnotes(path, ...)
}
\arguments{
\item{path}{(Character) The full path (including filename) to the Excel spreadsheet.}

\item{...}{Other named arguments that will be passed to \code{\link[readxl:read_excel]{readxl::read_excel()}}.}
}
\value{
A named list of class \code{field_notes}.
}
\description{
'Field notes' in this context is a spreadsheet that records the refinements
that a researcher makes throughout their interview process. This package is
opinionated about what these field notes should look like: use \code{\link[=create_field_notes_template]{create_field_notes_template()}}
to get a template for what the package accepts.
}
\examples{
\dontrun{
my_interviews <- import_coding_matrices("C:/path/to/folder/of/interviews")
my_codes <- score_codes(my_interviews)
my_notes <- import_field_notes("C:/path/to/notes.xlsx")

# Demonstrating usage in plots
plot_novelty(my_codes, refinements = my_notes)
plot_richness(my_codes, refinements = my_notes)
}

}
