% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_plot_novel_cumsum.R
\name{plot_novel_cumsum}
\alias{plot_novel_cumsum}
\title{Plot cumulative sum of novel codes over time}
\usage{
plot_novel_cumsum(score_df, refinements = integer(0))
}
\arguments{
\item{score_df}{(Dataframe) A dataframe of scored codes, as generated by \code{\link[=score_codes]{score_codes()}}.}

\item{refinements}{(Integer) A vector that lists when (in terms of interview
sequence) refinements were made to the interview questions. For example,
\code{c(10, 15)} means that interview questions were revised twice: First \strong{before}
the 10th interview, and then again \strong{before} the 15th interview.}
}
\value{
A ggplot object.
}
\description{
In this plot, the count of novel codes is cumulatively summed over all
interviews. Plateaus in this line could suggest a stoppin point for further
interviews, unless refinements to the interview questions introduce more
novel codes.
}
