% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_plot_novelty.R
\name{plot_novelty}
\alias{plot_novelty}
\title{Plot novelty of interviews over time}
\usage{
plot_novelty(score_df, refinements = integer(0))
}
\arguments{
\item{score_df}{(Dataframe) A dataframe of scored codes, as generated by \code{\link[=score_codes]{score_codes()}}.}

\item{refinements}{Either a list object generated by \code{\link[=import_field_notes]{import_field_notes()}},
or an Integer vector that lists when (in terms of interview sequence)
refinements were made to the interview questions. For example, \code{c(10, 15)}
means that interview questions were revised twice: First \strong{before} the
10th interview, and then again \strong{before} the 15th interview.}
}
\value{
A ggplot object.
}
\description{
Novel codes are information that has not been previously mentioned by other
interviewees. The trend of 'novel' interview codes provides insight into the
richness of qualitative information.

This plot shows the trend of novel code generation; in the most basic way,
the higher the number, the richer the information that has been generated in
the study. By showing novel codes in context with any refinements to the
questions, it also shows how that trend may have been affected by those
refinements, and whether novel code generation is plateauing.

This chart alone should not be used to decide on a stopping point because it
does not show the richness of individual interviews; some interviews are richer
than others, therefore consider also using \code{\link[=plot_richness]{plot_richness()}} to look at the
richness of each interview in terms of novel and duplicate codes.
}
\seealso{
\code{\link[=score_codes]{score_codes()}}, \code{\link[=import_field_notes]{import_field_notes()}}, \code{\link[=plot_richness]{plot_richness()}}, \code{\link[=save_last_plot]{save_last_plot()}}
}
