% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_plot_richness.R
\name{plot_richness}
\alias{plot_richness}
\title{Plot richness of interview codes over time}
\usage{
plot_richness(score_df, refinements = integer(0))
}
\arguments{
\item{score_df}{(Dataframe) A dataframe of scored codes, as generated by \code{\link[=score_codes]{score_codes()}}.}

\item{refinements}{Either a list object generated by \code{\link[=import_field_notes]{import_field_notes()}}, or
an Integer vector that lists when (in terms of interview
sequence) refinements were made to the interview questions. For example,
\code{c(10, 15)} means that interview questions were revised twice: First \strong{before}
the 10th interview, and then again \strong{before} the 15th interview.}
}
\value{
A ggplot object.
}
\description{
The full definition of novel and duplicate codes is in \code{\link[=score_codes]{score_codes()}}.
Briefly, 'novel' codes are topics/ideas/concepts that were not mentioned in
previous interviews, whereas 'duplicate' codes are topics that other
interviews have discussed previously.
}
\details{
This plot complements \code{\link[=plot_novelty]{plot_novelty()}} by visualising the richness of each
interview in terms of novel and duplicate codes, in context with any
refinements to interview questions that were made (marked by stars underneath
each bar). By examining this plot together with their field notes, researchers
can get insight into the effects of their refinements and the richness of the
data.
}
\seealso{
\code{\link[=score_codes]{score_codes()}}, \code{\link[=import_field_notes]{import_field_notes()}}, \code{\link[=plot_novelty]{plot_novelty()}}, \code{\link[=save_last_plot]{save_last_plot()}}
}
