% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prv_reshape_for_plots.R
\name{reshape_for_plots}
\alias{reshape_for_plots}
\title{Reshape score matrix for ggplot2.}
\usage{
reshape_for_plots(score_df, refinements = integer(0))
}
\arguments{
\item{score_df}{(Dataframe) A scored code matrix as generated by \code{\link[=score_codes]{score_codes()}}.}

\item{refinements}{Either a list object generated by \code{\link[=import_field_notes]{import_field_notes()}},
or an Integer vector that lists when (in terms of interview sequence)
refinements were made to the interview questions. For example, \code{c(10, 15)}
means that interview questions were revised twice: First \strong{before} the
10th interview, and then again \strong{before} the 15th interview.}
}
\value{
A dataframe.
}
\description{
Manually reshapes the data into Long format and adds a refinement group factor.
}
