% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_save_last_plot.R
\name{save_last_plot}
\alias{save_last_plot}
\title{Save the most recent plot to a file}
\usage{
save_last_plot(filename, size, dpi = 300, ...)
}
\arguments{
\item{filename}{(Character) The path and filename of the file to create.}

\item{size}{(Character) The output size of the file, in the form \verb{"width x height unit}. For example:
- \code{"5 x 7 in"}
- \code{"12 x 8 cm"}
- \code{"300 x 150 mm"}
- \code{"1920 x 1080 px"}}

\item{dpi}{(Integer) The resolution (dots per inch) of the output file.}

\item{...}{Other arguments passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\value{
This function returns nothing, but has the side-effect of writing a file to \code{filename}.
}
\description{
Save the most recent plot to a file
}
\examples{
\dontrun {
plot_novelty(interview_scores)

save_last_plot("my_file.png", "8 x 6 in")
}

}
