% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_plot_novelty.R
\name{plot_novelty}
\alias{plot_novelty}
\title{Plot novelty of interview codes over time}
\usage{
plot_novelty(
  score_df,
  refinements = integer(0),
  method = "lm",
  formula = y ~ poly(x, 8)
)
}
\arguments{
\item{score_df}{(Dataframe) A dataframe of scored codes, as generated by \code{\link[=score_codes]{score_codes()}}.}

\item{refinements}{(Integer) A vector that lists when (in terms of interview
sequence) refinements were made to the interview questions. For example,
\code{c(10, 15)} means that interview questions were revised twice: First \strong{before}
the 10th interview, and then again \strong{before} the 15th interview.}

\item{method}{(Character) Method for smoothing that will be passed to \code{\link[ggplot2:geom_smooth]{ggplot2::stat_smooth()}}.}

\item{formula}{(Formula) Formula for smoothing that will be passed to \code{\link[ggplot2:geom_smooth]{ggplot2::stat_smooth()}}.}
}
\value{
A ggplot object.
}
\description{
The 'novelty' of an interview is the proportion of codes that were mentioned
by this person, which were not previously mentioned by others.
}
